Moved points redux into separate files
