+++ app/javascript/packs/application/redux/points/index.js
@@ -1,40 +1,11 @@
-import api from "../../api";
-import { keyBy } from "lodash";
-
-export const CREATE_POINT = "CREATE_POINT";
-export const CREATE_POINT_SUCCESS = "CREATE_POINT_SUCCESS";
-export const CREATE_POINT_ERROR = "CREATE_POINT_ERROR";
-export const LIST_POINTS = "LIST_POINTS";
-export const LIST_POINTS_SUCCESS = "LIST_POINTS_SUCCESS";
-export const LIST_POINTS_ERROR = "LIST_POINTS_ERROR";
-export const GET_POINT = "GET_POINT";
-
-export const createPoint = (tourId, attributes) => async dispatch => {
-  dispatch({ type: CREATE_POINT });
-  const result = await api.createPoint(tourId, attributes);
-
-  if (result.ok) {
-    dispatch({
-      type: CREATE_POINT_SUCCESS,
-      data: { [response.data.id]: response.data }
-    });
-  } else {
-    dispatch({ type: CREATE_POINT_ERROR });
-  }
-
-  return response;
-};
-
-export const listPoints = tourId => async dispatch => {
-  dispatch({ type: LIST_POINTS });
-  const response = await api.listPoints(tourId);
-
-  if (response.ok) {
-    dispatch({ type: LIST_POINTS_SUCCESS, data: keyBy(response.data, "id") });
-  } else {
-    dispatch({ type: LIST_POINTS_ERROR });
-  }
-};
+import {
+  CREATE_POINT,
+  CREATE_POINT_SUCCESS,
+  CREATE_POINT_ERROR,
+  LIST_POINTS,
+  LIST_POINTS_SUCCESS,
+  LIST_POINTS_ERROR
+} from "./actions";
 
 export const INITIAL_STATE = {
   isFetching: false,
@@ -43,10 +14,17 @@ export const INITIAL_STATE = {
 
 export const reducer = (state = INITIAL_STATE, action) => {
   switch (action.type) {
+    case CREATE_POINT:
     case LIST_POINTS:
       return { ...state, isFetching: true };
+    case CREATE_POINT_SUCCESS:
     case LIST_POINTS_SUCCESS:
-      return { ...state, isFetching: false, data: action.data };
+      return {
+        ...state,
+        isFetching: false,
+        data: { ...state.data, ...action.data }
+      };
+    case CREATE_POINT_ERROR:
     case LIST_POINTS_ERROR:
       return { ...state, isFetching: false };
     default:
