+++ app/javascript/packs/application/redux/points/selectors.js
@@ -1,0 +1,7 @@
+export const getPoints = state => {
+  return Object.values(state.points.data);
+};
+
+export const isFetching = state => {
+  return state.points.isFetching;
+};
